op-deployer,ctb: migrate DeploySuperchain to new script API
