+++ op-deployer/pkg/deployer/bootstrap/superchain.go
@@ -197,8 +197,11 @@ func Superchain(ctx context.Context, cfg SuperchainConfig) (opcm.DeploySuperchai
 		return dso, fmt.Errorf("failed to create script host: %w", err)
 	}
 
-	dso, err = opcm.DeploySuperchain(
-		l1Host,
+	deploySuperchainScript, err := opcm.NewDeploySuperchainScript(l1Host)
+	if err != nil {
+		return dso, fmt.Errorf("failed to load DeploySuperchain script: %w", err)
+	}
+	dso, err = deploySuperchainScript.Run(
 		opcm.DeploySuperchainInput{
 			SuperchainProxyAdminOwner:  cfg.SuperchainProxyAdminOwner,
 			ProtocolVersionsOwner:      cfg.ProtocolVersionsOwner,
