+++ op-deployer/pkg/deployer/opcm/superchain_test.go
@@ -1,0 +1,42 @@
+package opcm
+
+import (
+	"math/big"
+	"testing"
+
+	"github.com/ethereum/go-ethereum/common"
+	"github.com/ethereum/go-ethereum/params"
+	"github.com/stretchr/testify/require"
+)
+
+func TestNewDeploySuperchainScript(t *testing.T) {
+	t.Run("should not fail with current version of DeploySuperchain contract", func(t *testing.T) {
+		// First we grab a test host
+		host1 := createTestHost(t)
+
+		// Then we load the script
+		//
+		// This would raise an error if the Go types didn't match the ABI
+		deploySuperchain, err := NewDeploySuperchainScript(host1)
+		require.NoError(t, err)
+
+		// Then we deploy
+		output, err := deploySuperchain.Run(DeploySuperchainInput{
+			Guardian:                   common.BigToAddress(big.NewInt(1)),
+			ProtocolVersionsOwner:      common.BigToAddress(big.NewInt(2)),
+			SuperchainProxyAdminOwner:  common.BigToAddress(big.NewInt(3)),
+			Paused:                     true,
+			RecommendedProtocolVersion: params.ProtocolVersion{1},
+			RequiredProtocolVersion:    params.ProtocolVersion{2},
+		})
+
+		// And do some simple asserts
+		require.NoError(t, err)
+		require.NotNil(t, output)
+		require.NotEqual(t, output.ProtocolVersionsImpl, common.Address{})
+		require.NotEqual(t, output.ProtocolVersionsProxy, common.Address{})
+		require.NotEqual(t, output.SuperchainConfigImpl, common.Address{})
+		require.NotEqual(t, output.SuperchainConfigProxy, common.Address{})
+		require.NotEqual(t, output.SuperchainProxyAdmin, common.Address{})
+	})
+}
